This should fix:
P2 Export Multiple to WAV: If user chooses to overwrite WAV files imported with preference set to "read directly", the exported files and the project are silenced. Mostly an issue if importing multiple files then splitting by track, but if one file is imported and a label is given the name of the imported file, that file, any following it on the timeline and the project itself will be silenced. Not actually a regression. Note: will not replicate if using the "FFmpeg-compatible files" filter, because that always makes a copy of the data irrespective of the preference setting.