Fix (I hope) for P2 "Chains: EQ applies parameters last used via Effect menu instead of those stored in chain."
Remembered settings should now be largely separate for the batch-chain and
effect-menu uses of EQ.

When editing parameters for an effect in a chain, the parameters dialog now has
the command selection dialog as its parent (rather than NULL). This means it
appears on top & centered properly, and also allows the EQ effect to detect
whether it is being edited as part of a chain (little bit of a hack). If part
of a chain, the EQ dialog checks whether the custom curve was selected when OK
was pressed - if so, a message is displayed which explains that the curve must
have a distinct name in order to be used in a batch chain.

To speed up testing, I also added a 'ChainName' parameter to BatchEvalCommand
so that saved chains can be used from a script. (Not likely to be used much as
the script can do everything the chains do, but it doesn't hurt).