Fix crash at Audacity termination if user clicks the window close button (X).  Dialog would be deleted, but wxWidgets didn't know about it so it tried to get rid of it again at termination.
And changes the way bundles are loaded on OSX to only use CoreFoundation to get the executable path and the use the dl...() functions to actually load the bundle.  This was done since some of the CoreFoundation functions would cause the process to crash if the bundle couldn't be loaded due to missing symbols and such.