I accidentially included the patch for bug #647 along with the patch for bug #115.
But, since there's some definite preferences on which is correct, I've
left the change in, but added a simple way (SNAP_TO_NEAREST define in Snap.c)
to change it once a decision is made.

If set SNAP_TO_NEAREST back to true since that is the current (as of 2.0.5)
behavior.