Further Nyquist Prompt bug fixing from Steve
With the addition of the preview button, it became evident
that if an effect were to return text instead of audio, the
processing loop would not stop properly.

And he also found a case where the debug state wasn't reset
after using the debug button.  If this was following by the
preview button the next time the effect was used, the debug
window would reappear before the preview.