Remove conditional build of DirectX support
With the move to VS2013, we no longer need the DirectX SDK
and, therefore, we can always enable DirectX support.

If an individual user doesn't want DirectX support, they can
always disable it for their private copy.  But, I believe we
will always release with DX support, so it is now on always.

For those that have the DX SDK installed just for building
Audacity, you may now remove it and you should still get
DX support in Audacity.