Restore default meter size.
The meters are back to the length they used to be (with a little space lost for the grabber).  If we want to increase their default lengths then we should look at the space saver idea (make them less tall) or else widen the initial size of Audacity main screen.