Cleanup chunk setting and preset saving/loading
When looking into the +morphfilter issue, I noticed that
I wasn't consistently handling preset chunks.  While I
didn't have any effects complaining about it, some of it
was (near as I can deduce...this stuff isn't well documented)
probably wrong.

So, I cleaned it all up and in the process added more validity
checking when loading presets for all 3 preset formats.

Incidentially, the +morphfilter still crashes Audacity...but
it also crashes Cubase, so we're in good company.  :-)  I've
sent an email to the author to see if I'm setting the chunks
correctly or he has a known issue.